Fix norme of libft

+++ Libft/ft_strcmp.c
@@ -23,8 +23,7 @@ int				ft_strcmp(const char *str1, const char *str2)
 		return (*str2);
 	if (str2 == NULL)
 		return (*str1);
-	while ((unsigned int)(str1[i]) &&
-			(unsigned int)(str1[i]) == (unsigned int)(str2[i]))
+	while (str1[i] && str1[i] == str2[i])
 		i++;
 	return ((unsigned char)str1[i] - (unsigned char)str2[i]);
 }
@@ -40,8 +39,7 @@ unsigned int	ft_strlcmp(const char *str1, const char *str2)
 		return (*str2);
 	if (str2 == NULL)
 		return (*str1);
-	while ((unsigned int)(str1[i]) &&
-			(unsigned int)(str1[i]) == (unsigned int)(str2[i]))
+	while (str1[i] && str1[i] == str2[i])
 		i++;
 	if (str2[i] == 0)
 		return (i);
